requirements.txt
# Deep learning core (torch>=2.1 is the first release with stable MPS support)
torch>=2.1
torchvision>=0.16

# Vision-Language Models
open_clip_torch>=2.24
transformers>=4.36

# Classical ML & fairness
scikit-learn>=1.3
fairlearn>=0.9

# Explainability
grad-cam>=1.5
captum>=0.6

# Plotting & tables
matplotlib>=3.7
seaborn>=0.13
pandas>=2.0
numpy>=1.24
pillow>=10.0

# Utilities
pyyaml>=6.0
tqdm>=4.65
tabulate>=0.9          # for DataFrame.to_markdown(...)
jinja2>=3.1.2          # for DataFrame.to_latex(...)

# Dataset download helpers
kagglehub
gdown
requests
